refactor(l2): remove references to vm internal api. (#2299)
**Motivation**
L2 code was accessing internal apis from the vm crate, specifically
`revm` constructs. This is attempt to replace those with the public api,
so that we can easily switch between revm and levm.

**Description**
- Replaces references to `ethrex_vm::backends::` from the prover
backends.
- Moved `ExecutionDB ` to `vm/db.rs`. It is still somewhat coupled with
revm but less than before. It should be totally decoupled.

---------

+++ crates/vm/levm/bench/revm_comparison/src/benchmark.rs
@@ -20,7 +20,7 @@ fn main() {
 
     let benchmark = std::env::args().nth(2).expect(usage);
 
-    let runs: usize = std::env::args()
+    let runs: u64 = std::env::args()
         .nth(3)
         .unwrap_or_else(|| "10".to_string()) // Default to 10 runs
         .parse()
